requirements.txt
# FastAPI & server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.6

# GPS / geospatial
gpxpy==1.6.2
shapely==2.0.2
geopy==2.4.1

# Auth & validation
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9

# Config & env
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Storage (S3-compatible)
boto3==1.34.34

# Utils
python-dateutil==2.8.2
numpy==1.26.4
